service_airflow/requirements.txt
apache-airflow-providers-celery==3.8.5
apache-airflow-providers-postgres
apache-airflow-providers-docker
apache-airflow-providers-sftp
apache-airflow-providers-grpc
psycopg[binary]
Faker